+++ src/constants/fen.js
@@ -1,2 +1,0 @@
-export const INITIAL_POSITION =
- 'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1';