+++ src/constants/index.js
@@ -1,0 +1,31 @@
+export const INITIAL_POSITION_FEN =
+  'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1';
+ 
+ export const INITIAL_POSITION =
+  'rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR';
+ 
+ export const INITIAL_POSITION_ARRAY_OF_PIECES = [
+  "r", "n", "b", "q", "k", "b", "n", "r",
+  "p", "p", "p", "p", "p", "p", "p", "p",
+  " ", " ", " ", " ", " ", " ", " ", " ",
+  " ", " ", " ", " ", " ", " ", " ", " ",
+  " ", " ", " ", " ", " ", " ", " ", " ",
+  " ", " ", " ", " ", " ", " ", " ", " ",
+  "P", "P", "P", "P", "P", "P", "P", "P",
+  "R", "N", "B", "Q", "K", "B", "N", "R"
+ ];
+ 
+ export const TEST_POSITION =
+   '3r2k1/p6p/7b/8/N1Pp4/P2K1p2/4r3/R2R4';
+
+export const TEST_POSITION_ARRAY_OF_PIECES = [
+  " ", " ", " ", "r", " ", " ", "k", " ",
+  "p", " ", " ", " ", " ", " ", " ", "p",
+  " ", " ", " ", " ", " ", " ", " ", "b",
+  " ", " ", " ", " ", " ", " ", " ", " ",
+  "N", " ", "P", "p", " ", " ", " ", " ",
+  "P", " ", " ", "K", " ", "p", " ", " ",
+  " ", " ", " ", " ", "r", " ", " ", " ",
+  "R", " ", " ", "R", " ", " ", " ", " "
+];
+